backend/pyproject.toml
[project]
name = "oopsnote-backend"
version = "0.1.0"
description = "FastAPI orchestration service for OopsNote"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "pydantic>=2.9.0",
  "openai>=1.55.0",
  "Pillow>=10.0.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["storage*", "tests*"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pylint.messages_control]
# Ignore docstring requirements for this codebase.
disable = [
  # Style/doc guidelines we don't enforce.
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",

  # Allow pragmatic error handling and interface-compatible signatures.
  "broad-exception-caught",
  "unused-argument",
]
